# 1.4.0-a - Net10 support and update nuget packages - Upgrade mapping profiles - ExpirationOptions to enable/disable timer and intervals - SemaphoreOptions for locking intervals - Cosmos context changes

+++ src/V1/ServiceBricks.Cache/Background/CacheExpirationTask.cs
@@ -64,6 +64,10 @@ namespace ServiceBricks.Cache
 
                 foreach (var item in respExpired.Item.List)
                 {
+                    // Cancel if requested
+                    if (cancellationToken.IsCancellationRequested)
+                        return;
+
                     // AI: Delete the expired cache data.
                     await _dataApiService.DeleteAsync(item.StorageKey);
                 }
